+++ tts_tradeapi.cpp
@@ -4,12 +4,13 @@
 #include "json.hpp"
 #include "tts_common.h"
 #include "tts_dll.h"
+#include "tts_activeclients.h"
 #include <QTextCodec>
 
 using namespace std;
 using json = nlohmann::json;
 
-#define INVALID_CLIENT_ID "无效的clientID或者Trade.DLL无效"
+#define INVALID_CLIENT_ID "无效的clientId或者Trade.DLL无效"
 
 TTS_TradeApi::TTS_TradeApi(const TTS_SettingObject& so):outputUtf8(true),_so(so)
 {
@@ -43,84 +44,133 @@ void TTS_TradeApi::setOutputUtf8(bool utf8) {
  * @return data -> { "client_id" : xxx }
  *
  */
-json TTS_TradeApi::logon(const char* IP, const short Port,
-                        const char* Version, short YybID,
-                        const char* AccountNo, const char* TradeAccount,
-                        const char* JyPassword, const char* TxPassword) {
-    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, AccountNo);
-    json j = dll->logon(IP, Port, Version, YybID, AccountNo, TradeAccount, JyPassword, TxPassword);
+json TTS_TradeApi::logon(const char* ip, const short port,
+                        const char* version, short yybId,
+                        const char* accountNo, const char* tradeAccount,
+                        const char* jyPassword, const char* txPassword) {
+    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, accountNo);
+    json j = dll->logon(ip, port, version, yybId, accountNo, tradeAccount, jyPassword, txPassword);
     if (j[TTS_SUCCESS].get<bool>() == true) {
-        uint32_t loginId = j[TTS_DATA]["client_id"].get<uint32_t>();
-        j[TTS_DATA]["client_id"] = genSessionId(dll->getSeq(), loginId);
+        uint32_t clientId = j[TTS_DATA]["client_id"].get<uint32_t>();
+        uint32_t sessionId = genSessionId(dll->getSeq(), clientId);
+        j[TTS_DATA]["client_id"] = sessionId;
+        // 记录登陆行为
+        TTS_ActiveClients::ins()->addNewEntry(
+                    sessionId,
+                    clientId,
+                    ip,
+                    port,
+                    version,
+                    yybId,
+                    accountNo,
+                    tradeAccount
+                    );
     }
     return j;
 }
 
 /**
  * @brief TTS_TradeApi::logoff 登出
- * @param ClientID cilent_id
+ * @param clientId cilent_id
  * @return success => true/false
  */
-json TTS_TradeApi::logoff(int ClientID) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::logoff(int clientId) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
     json j = dll->logoff(loginId);
+    // 如果登出成功
+    if (j["success"].get<bool>()) {
+        TTS_ActiveClients::ins()->removeEntryBySessionId(clientId);
+    }
+
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
 /**
  * @brief TTS_TradeApi::queryData 查询信息
- * @param ClientID client_id
+ * @param clientId client_id
  * @param Category 信息类别
  * @return [{}, {}, {} ]
  */
-json TTS_TradeApi::queryData(int ClientID, int Category) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::queryData(int clientId, int category) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->queryData(loginId, Category);
+    json j = dll->queryData(loginId, category);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
 
-json TTS_TradeApi::sendOrder(int ClientID, int Category ,int PriceType, const char* Gddm,  const char* Zqdm , float Price, int Quantity) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::sendOrder(int clientId, int category ,int priceType, const char* gddm,  const char* zqdm , float price, int quantity) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->sendOrder(loginId, Category, PriceType, Gddm, Zqdm, Price, Quantity);
+    json j = dll->sendOrder(loginId, category, priceType, gddm, zqdm, price, quantity);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
-json TTS_TradeApi::cancelOrder(int ClientID, const char *ExchangeID, const char *hth) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::cancelOrder(int clientId, const char *exchangeID, const char *hth) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->cancelOrder(loginId, ExchangeID, hth);
+    json j = dll->cancelOrder(loginId, exchangeID, hth);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
-json TTS_TradeApi::getQuote(int ClientID, const char *Zqdm) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::getQuote(int clientId, const char *zqdm) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->getQuote(loginId, Zqdm);
+    json j = dll->getQuote(loginId, zqdm);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
-json TTS_TradeApi::repay(int ClientID, const char *Amount) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::repay(int clientId, const char *amount) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->repay(loginId, Amount);
+    json j = dll->repay(loginId, amount);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }
 
-json TTS_TradeApi::queryHistoryData(int ClientID, int Category, const char* BeginDate, const char* EndDate) {
-    int loginId = getLoginIdBySessionId(ClientID);
-    int seq = getSeqBySessionId(ClientID);
+json TTS_TradeApi::queryHistoryData(int clientId, int category, const char* beginDate, const char* endDate) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
     std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
-    json j = dll->queryHistoryData(loginId, Category, BeginDate, EndDate);
+    json j = dll->queryHistoryData(loginId, category, beginDate, endDate);
+    return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
+}
+
+json TTS_TradeApi::queryDatas(int clientId, int categories[], int count) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
+    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
+    json j = dll->queryDatas(loginId, categories, count);
+    return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
+}
+
+json TTS_TradeApi::sendOrders(int clientId, int categories[], int priceTypes[], const char *gddms[], const char *zqdms[], float prices[], int quantities[], int count) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
+    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
+    json j = dll->sendOrders(loginId, categories, priceTypes, gddms, zqdms, prices, quantities, count);
+    return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
+}
+
+json TTS_TradeApi::cancelOrders(int clientId, const char *exchangeIds[], const char *hths[], int count) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
+    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
+    json j = dll->cancelOrders(loginId, exchangeIds, hths, count);
+    return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
+}
+
+json TTS_TradeApi::getQuotes(int clientId, const char *zqdms[], int count) {
+    int loginId = getLoginIdBySessionId(clientId);
+    int seq = getSeqBySessionId(clientId);
+    std::shared_ptr<TTS_Dll> dll = TTS_Dll::getInstance(_so, seq);
+    json j = dll->getQuotes(loginId, zqdms, count);
     return (dll == nullptr) ? jsonError(INVALID_CLIENT_ID): j;
 }